Split off some functions into drm-util
